+++ app/src/main/java/com/munscore/DayThreeScore.java
@@ -78,22 +78,22 @@ public class DayThreeScore extends AppCompatActivity {
                             Intent i = getIntent();
                             String type = i.getStringExtra("type");
                             if(type.equals("poo")){
-                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_ORDER, 3, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_ORDER, 3, Float.parseFloat(c[0]));
                             }
                             if(type.equals("poi")){
-                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_INFO, 3, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_INFO, 3, Float.parseFloat(c[0]));
                             }
                             if(type.equals("chit")){
-                                mydb.InsertTabScore(getApplicationContext(), name, CHIT, 3, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, CHIT, 3, Float.parseFloat(c[0]));
                             }
                             if(type.equals("speech")){
                                 mydb.insertScore(getApplicationContext(), cr, c, 3, name);
                             }
                             if(type.equals("dr")){
-                                mydb.InsertTabScore(getApplicationContext(), name, DR, 3, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, DR, 3, Float.parseFloat(c[0]));
                             }
                             if(type.equals("dir")){
-                                mydb.InsertTabScore(getApplicationContext(), name, DIRECTIVE, 3, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, DIRECTIVE, 3, Float.parseFloat(c[0]));
                             }
                             Intent in = new Intent(getApplicationContext(), CountryDetails.class);
                             setResult(RESULT_OK, in);
