+++ app/src/main/java/com/munscore/DayTwoScore.java
@@ -78,22 +78,22 @@ public class DayTwoScore extends AppCompatActivity{
                             Intent i = getIntent();
                             String type = i.getStringExtra("type");
                             if(type.equals("poo")){
-                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_ORDER, 2, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_ORDER, 2, Float.parseFloat(c[0]));
                             }
                             if(type.equals("poi")){
-                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_INFO, 2, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, POINT_OF_INFO, 2, Float.parseFloat(c[0]));
                             }
                             if(type.equals("chit")){
-                                mydb.InsertTabScore(getApplicationContext(), name, CHIT, 2, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, CHIT, 2, Float.parseFloat(c[0]));
                             }
                             if(type.equals("speech")){
                                 mydb.insertScore(getApplicationContext(), cr, c, 2, name);
                             }
                             if(type.equals("dr")){
-                                mydb.InsertTabScore(getApplicationContext(), name, DR, 2, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, DR, 2, Float.parseFloat(c[0]));
                             }
                             if(type.equals("dir")){
-                                mydb.InsertTabScore(getApplicationContext(), name, DIRECTIVE, 2, Integer.parseInt(c[0]));
+                                mydb.InsertTabScore(getApplicationContext(), name, DIRECTIVE, 2, Float.parseFloat(c[0]));
                             }
                             Intent in = new Intent(getApplicationContext(), CountryDetails.class);
                             setResult(RESULT_OK, in);
